+++ src/components/ProductList/ProductListResults.jsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import { ProductListStyle } from "./ProductList";
+import ProductListStyle from "./ProductListStyle";
 import ProductCard from "../ProductCard";
 import Button from "../ui/Button";
